refactor: 메인 화면 옵션 클래스 위치 변경
